backend/apps/products/tests/factories.py
import factory
from decimal import Decimal
from apps.products.models import Category, Product


class CategoryFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Category

    name = factory.Sequence(lambda n: f'Category {n}')
    slug = factory.Sequence(lambda n: f'category-{n}')
    is_active = True


class ProductFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Product

    name = factory.Sequence(lambda n: f'Product {n}')
    slug = factory.Sequence(lambda n: f'product-{n}')
    description = 'A test product.'
    price = Decimal('29.99')
    stock = 50
    category = factory.SubFactory(CategoryFactory)
    is_active = True
